+++ src/NodaMoney/Context/CashDenominationRounding.cs
@@ -10,5 +10,5 @@ internal record CashDenominationRounding : IRoundingStrategy
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public decimal Round(decimal amount, CurrencyInfo currencyInfo, int? decimals) => throw new NotImplementedException();
+    public decimal Round(decimal amount, Currency currency, int? decimals) => throw new NotImplementedException();
 }